pyproject.toml
[project]
name = "medmnist-classification"
version = "0.1.0"
description = ""
authors = [
    {name = "Sam Winebrake",email = "samwinebrake@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "medmnist (>=3.0.2,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "timm (>=1.0.26,<2.0.0)",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pillow (>=12.2.0,<13.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]
